+++ plugins/gh-datainmap/CHANGELOG.md
@@ -1,3 +1,5 @@
+- Filters <aside> voorzien van een label.
+
 **1.6.5 (2020-07-21)**
 - Filters gegroepeerd (role=group) voor toegankelijkheid.